deprecate empty cache views in favor of new noop implementation
fix bug in SerialRequester

make CustomizableEntityCache noop instead of throwing on unimplemented methods

add missing jsr305 annotations to CacheView
