+++ src/main/java/com/mewna/catnip/cache/CustomizableEntityCache.java
@@ -39,10 +39,12 @@ import com.mewna.catnip.entity.misc.Emoji.CustomEmoji;
 import com.mewna.catnip.entity.user.Presence;
 import com.mewna.catnip.entity.user.User;
 import com.mewna.catnip.entity.user.VoiceState;
+import com.mewna.catnip.util.SafeVertxCompletableFuture;
 import io.vertx.core.Future;
 import io.vertx.core.json.JsonObject;
 
 import javax.annotation.Nonnull;
+import javax.annotation.OverridingMethodsMustInvokeSuper;
 import java.util.Collection;
 import java.util.Map;
 import java.util.concurrent.CompletableFuture;
@@ -50,207 +52,211 @@ import java.util.concurrent.CompletableFuture;
 /**
  * A customizable entity cache is basically just what it sounds like - a cache
  * whose behaviour is ENTIRELY meant to be customized. By default, this cache
- * will throw an {@link UnsupportedOperationException} on all methods it does
- * not implement. If you want to write a custom partial cache, this is probably
+ * will return {@link CacheView#noop() noop caches} and futures completed with
+ * {@code null}. If you want to write a custom partial cache, this is probably
  * the best place to start so that you don't have lots of empty method bodies.
  *
  * @author amy
  * @since 3/7/19.
  */
 public abstract class CustomizableEntityCache implements EntityCacheWorker {
+    protected Catnip catnip;
+    
+    @OverridingMethodsMustInvokeSuper
+    @Nonnull
+    @Override
+    public EntityCache catnip(@Nonnull final Catnip catnip) {
+        this.catnip = catnip;
+        return this;
+    }
+    
     @Nonnull
     @Override
     public Future<Void> updateCache(@Nonnull final String eventType, final int shardId, @Nonnull final JsonObject payload) {
-        throw new UnsupportedOperationException("Not implemented.");
+        return Future.succeededFuture();
     }
     
     @Override
     public void bulkCacheUsers(final int shardId, @Nonnull final Collection<User> users) {
-        throw new UnsupportedOperationException("Not implemented.");
+        //noop
     }
     
     @Override
     public void bulkCacheChannels(final int shardId, @Nonnull final Collection<GuildChannel> channels) {
-        throw new UnsupportedOperationException("Not implemented.");
+        //noop
     }
     
     @Override
     public void bulkCacheRoles(final int shardId, @Nonnull final Collection<Role> roles) {
-        throw new UnsupportedOperationException("Not implemented.");
+        //noop
     }
     
     @Override
     public void bulkCacheMembers(final int shardId, @Nonnull final Collection<Member> members) {
-        throw new UnsupportedOperationException("Not implemented.");
+        //noop
     }
     
     @Override
     public void bulkCacheEmoji(final int shardId, @Nonnull final Collection<CustomEmoji> emoji) {
-        throw new UnsupportedOperationException("Not implemented.");
+        //noop
     }
     
     @Override
     public void bulkCachePresences(final int shardId, @Nonnull final Map<String, Presence> presences) {
-        throw new UnsupportedOperationException("Not implemented.");
+        //noop
     }
     
     @Override
     public void bulkCacheVoiceStates(final int shardId, @Nonnull final Collection<VoiceState> voiceStates) {
-        throw new UnsupportedOperationException("Not implemented.");
+        //noop
     }
     
     @Override
     public void invalidateShard(final int id) {
-        throw new UnsupportedOperationException("Not implemented.");
-    }
-    
-    @Nonnull
-    @Override
-    public EntityCache catnip(@Nonnull final Catnip catnip) {
-        throw new UnsupportedOperationException("Not implemented.");
+        //noop
     }
     
     @Nonnull
     @Override
     public CompletableFuture<Guild> guildAsync(final long id) {
-        throw new UnsupportedOperationException("Not implemented.");
+        return SafeVertxCompletableFuture.completedFuture(catnip, null);
     }
     
     @Nonnull
     @Override
     public NamedCacheView<Guild> guilds() {
-        throw new UnsupportedOperationException("Not implemented.");
+        return CacheView.noop();
     }
     
     @Nonnull
     @Override
     public CompletableFuture<User> userAsync(final long id) {
-        throw new UnsupportedOperationException("Not implemented.");
+        return SafeVertxCompletableFuture.completedFuture(catnip, null);
     }
     
     @Nonnull
     @Override
     public NamedCacheView<User> users() {
-        throw new UnsupportedOperationException("Not implemented.");
+        return CacheView.noop();
     }
     
     @Nonnull
     @Override
     public CompletableFuture<Presence> presenceAsync(final long id) {
-        throw new UnsupportedOperationException("Not implemented.");
+        return SafeVertxCompletableFuture.completedFuture(catnip, null);
     }
     
     @Nonnull
     @Override
     public CacheView<Presence> presences() {
-        throw new UnsupportedOperationException("Not implemented.");
+        return CacheView.noop();
     }
     
     @Nonnull
     @Override
     public CompletableFuture<Member> memberAsync(final long guildId, final long id) {
-        throw new UnsupportedOperationException("Not implemented.");
+        return SafeVertxCompletableFuture.completedFuture(catnip, null);
     }
     
     @Nonnull
     @Override
     public NamedCacheView<Member> members(final long guildId) {
-        throw new UnsupportedOperationException("Not implemented.");
+        return CacheView.noop();
     }
     
     @Nonnull
     @Override
     public NamedCacheView<Member> members() {
-        throw new UnsupportedOperationException("Not implemented.");
+        return CacheView.noop();
     }
     
     @Nonnull
     @Override
     public CompletableFuture<Role> roleAsync(final long guildId, final long id) {
-        throw new UnsupportedOperationException("Not implemented.");
+        return SafeVertxCompletableFuture.completedFuture(catnip, null);
     }
     
     @Nonnull
     @Override
     public NamedCacheView<Role> roles(final long guildId) {
-        throw new UnsupportedOperationException("Not implemented.");
+        return CacheView.noop();
     }
     
     @Nonnull
     @Override
     public NamedCacheView<Role> roles() {
-        throw new UnsupportedOperationException("Not implemented.");
+        return CacheView.noop();
     }
     
     @Nonnull
     @Override
     public CompletableFuture<GuildChannel> channelAsync(final long guildId, final long id) {
-        throw new UnsupportedOperationException("Not implemented.");
+        return SafeVertxCompletableFuture.completedFuture(catnip, null);
     }
     
     @Nonnull
     @Override
     public NamedCacheView<GuildChannel> channels(final long guildId) {
-        throw new UnsupportedOperationException("Not implemented.");
+        return CacheView.noop();
     }
     
     @Nonnull
     @Override
     public NamedCacheView<GuildChannel> channels() {
-        throw new UnsupportedOperationException("Not implemented.");
+        return CacheView.noop();
     }
     
     @Nonnull
     @Override
     public CompletableFuture<UserDMChannel> dmChannelAsync(final long id) {
-        throw new UnsupportedOperationException("Not implemented.");
+        return SafeVertxCompletableFuture.completedFuture(catnip, null);
     }
     
     @Nonnull
     @Override
     public CacheView<UserDMChannel> dmChannels() {
-        throw new UnsupportedOperationException("Not implemented.");
+        return CacheView.noop();
     }
     
     @Nonnull
     @Override
     public CompletableFuture<CustomEmoji> emojiAsync(final long guildId, final long id) {
-        throw new UnsupportedOperationException("Not implemented.");
+        return SafeVertxCompletableFuture.completedFuture(catnip, null);
     }
     
     @Nonnull
     @Override
     public NamedCacheView<CustomEmoji> emojis(final long guildId) {
-        throw new UnsupportedOperationException("Not implemented.");
+        return CacheView.noop();
     }
     
     @Nonnull
     @Override
     public NamedCacheView<CustomEmoji> emojis() {
-        throw new UnsupportedOperationException("Not implemented.");
+        return CacheView.noop();
     }
     
     @Nonnull
     @Override
     public CompletableFuture<VoiceState> voiceStateAsync(final long guildId, final long id) {
-        throw new UnsupportedOperationException("Not implemented.");
+        return SafeVertxCompletableFuture.completedFuture(catnip, null);
     }
     
     @Nonnull
     @Override
     public CacheView<VoiceState> voiceStates(final long guildId) {
-        throw new UnsupportedOperationException("Not implemented.");
+        return CacheView.noop();
     }
     
     @Nonnull
     @Override
     public CacheView<VoiceState> voiceStates() {
-        throw new UnsupportedOperationException("Not implemented.");
+        return CacheView.noop();
     }
     
     @Nonnull
     @Override
     public CompletableFuture<User> selfUserAsync() {
-        throw new UnsupportedOperationException("Not implemented.");
+        return SafeVertxCompletableFuture.completedFuture(catnip, null);
     }
 }
