+++ src/main/java/com/mewna/catnip/cache/view/EmptyNamedCacheView.java
@@ -27,67 +27,19 @@
 
 package com.mewna.catnip.cache.view;
 
-import javax.annotation.Nonnull;
-import java.util.Collection;
-import java.util.Collections;
-import java.util.Objects;
-import java.util.function.Supplier;
-
 /**
  * Always empty {@link NamedCacheView named cache view}.
  *
  * @author natanbc
+ *
  * @see NamedCacheView#empty()
+ *
  * @since 12/15/18
+ *
+ * @deprecated Use {@link CacheView#noop()}.
  */
-public class EmptyNamedCacheView<T> extends EmptyCacheView<T> implements NamedCacheView<T> {
+@Deprecated
+public class EmptyNamedCacheView<T> extends EmptyCacheView<T> {
+    @Deprecated
     public static final NamedCacheView<?> INSTANCE = new EmptyNamedCacheView<>();
-    
-    @Nonnull
-    @Override
-    public Collection<T> findByName(@Nonnull final String name, final boolean ignoreCase) {
-        return Collections.emptyList();
-    }
-    
-    @Nonnull
-    @Override
-    public <C extends Collection<T>> C findByName(@Nonnull final String name, final boolean ignoreCase, @Nonnull final Supplier<C> supplier) {
-        return Objects.requireNonNull(supplier.get(), "Provided collection may not be null");
-    }
-    
-    @Nonnull
-    @Override
-    public Collection<T> findByNameContains(@Nonnull final String name, final boolean ignoreCase) {
-        return Collections.emptyList();
-    }
-    
-    @Nonnull
-    @Override
-    public <C extends Collection<T>> C findByNameContains(@Nonnull final String name, final boolean ignoreCase, @Nonnull final Supplier<C> supplier) {
-        return Objects.requireNonNull(supplier.get(), "Provided collection may not be null");
-    }
-    
-    @Nonnull
-    @Override
-    public Collection<T> findByNameStartsWith(@Nonnull final String name, final boolean ignoreCase) {
-        return Collections.emptyList();
-    }
-    
-    @Nonnull
-    @Override
-    public <C extends Collection<T>> C findByNameStartsWith(@Nonnull final String name, final boolean ignoreCase, @Nonnull final Supplier<C> supplier) {
-        return Objects.requireNonNull(supplier.get(), "Provided collection may not be null");
-    }
-    
-    @Nonnull
-    @Override
-    public Collection<T> findByNameEndsWith(@Nonnull final String name, final boolean ignoreCase) {
-        return Collections.emptyList();
-    }
-    
-    @Nonnull
-    @Override
-    public <C extends Collection<T>> C findByNameEndsWith(@Nonnull final String name, final boolean ignoreCase, @Nonnull final Supplier<C> supplier) {
-        return Objects.requireNonNull(supplier.get(), "Provided collection may not be null");
-    }
 }
